requirements.txt
# 데이터 수집
yfinance==0.2.38
praw==7.7.1
feedparser==6.0.11
finnhub-python==2.4.20

# NLP / ML
transformers==4.40.0
torch==2.3.0
xgboost==2.0.3
scikit-learn==1.4.2
shap==0.45.0

# 데이터 처리
pandas==2.2.2
numpy==1.26.4
ta==0.11.0

# MLOps
clearml==1.16.0
evidently==0.4.22
codecarbon==2.3.4

# API / 서버
fastapi==0.111.0
uvicorn==0.30.0

# 기타
python-dotenv==1.0.1
boto3==1.34.0